requirements.txt
# Core Flask Framework
Flask==3.1.1
Flask-SocketIO==5.5.1
flask-cors==6.0.1
Flask-Compress==1.15
flask_compress
python-dotenv==1.0.1

# ASR Backend Dependencies
groq==0.31.0
# For NVIDIA NeMo (local): install the following
torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
nemo-toolkit[asr]>=1.20.0
librosa>=0.10.0
soundfile>=0.12.0

# Additional Utilities
ffmpeg-python==0.2.0
python-Levenshtein==0.27.1

# Server
gunicorn
eventlet==0.33.3
